Add test certification key, resolves #1
The 'Jesse Cox for Windows_TemporaryKey.pfx' file was added to make it easier for cloning and testing this application. Without it, users previously had to generate a new test certificate, find their publisher ID and ensure it matched both the certificate and the Package.appxmanifest file.

Store certification process will replace the temp keys and publisher ID with the user's true publisher ID, so there's no privacy risk in including this file in a public repo.